refactor: share _document across sites (#1637)
